requirements.txt
google-generativeai
PyMuPDF
Pillow
python-dotenv
opencc-python-reimplemented
openai